Fixes to printing to trace file. You should now see if the location was an intruction fetch, data write, and data read.

+++ src/pdp11simcontroller.cpp
@@ -45,13 +45,16 @@ PDP11SimController::~PDP11SimController()
 
 void PDP11SimController::run()
 {
+	//First instruction fetch before while loop is for check for HALT
 	fetch();
 	while(ci != HALT_OPCODE)
 	{
+		//Second instruction fetch is to fetch the next instruction from memory
 		fetch();
 
 		//Print to the trace file (instruction fetch)
 		PRINT_TO_FILE(pc.getVal(), 2);
+
 		if (ci != NOP_OPCODE) 
 		{
 			if (!decode())
@@ -2035,7 +2038,7 @@ bool PDP11SimController::PRINT_TO_FILE(OctalWord address, char type)
 		traceFile.open("trace_file.txt", ofstream::out | ofstream::app);
 
 		//Write to the end of the file
-		traceFile << type << " " << tempAddr << endl;
+		traceFile << (int)type << " " << tempAddr << endl;
 	}
 	catch (exception e)
 	{
